requirements.txt
pytest-watch==4.1.0
pytest-django==3.1.2
pytest-coverage==0.0
pytest-benchmark==3.1.1
pep8==1.7.0
markdown==2.6.9
ipdb==0.10.3
flake8==3.4.1
djangorestframework==3.7.0
django==1.11.6
django-cors-headers==2.1.0
chance==0.104
